refactor: improve chembl utilities
