+++ get_input_initialisation.py
@@ -70,9 +70,6 @@ def run(args: argparse.Namespace) -> int:
     except KeyError as exc:
         logger.error("required table '%s' missing", exc.args[0])
         return 1
-    except Exception as exc:  # pragma: no cover - defensive
-        logger.error("%s", exc)
-        return 1
 
 
 def build_parser() -> argparse.ArgumentParser:
