+++ requirements.txt
@@ -7,3 +7,4 @@ ruff>=0.0.290
 mypy>=1.8
 
 openpyxl>=3.1
+pyarrow>=12.0
